[UPDATE] Add diff-so-fancy
Updated configuration added using brew. But also created script in shell-setup to fetch from repo. Note this script saves it to a local user bin folder, which needs to be added to the path. This also adds the recommended global git configuration changes to update the colors for better readability and set it as a default diff pager