[UPDATE] Add alias for youtube-dl and google suggstions
Youtube-dl. Installed using brew install youtube-dl. Inspired by Joses's alias file we have added youtube-dl and his associated, flags and aliases to download audio only. This would be useful to re-listen to things I've already listend and store in a conveniently accessible location, on my mac devices.

Google autosuggestions. This is an interesting function, not sure how he uses it. But it fetches the top 10 autocompletes suggested by google. This requires installation of JQ (brew install JQ), which is a SED like tool for JSON input.