[UPDATE] Source fzf file
On Linux installation it seems to automatically add a line to the zshrc file, that sources the required scripts for fzf, on mac this does not happen. Here we are storing it in the zprofile file to keep the zshrc file clean. However can just keep it in the zshrc file in the future